requirements.txt
# Wildfire Prediction Project Requirements
# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning
scikit-learn>=1.0.0

# Progress bars
tqdm>=4.62.0

# Image processing
Pillow>=8.3.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
streamlit>=1.0.0

